Added a test for configuration loading.
Getting this test to pass required some fixes to both the code, as well
as the test job files.

An addition was the ChildProcessSkeleton class, which allows a
ChildProcess to be configured without having an event queue, but which
will not allow the process to be started unless the event queue is set.