Fixed the daemon class to kill the parent.
Although I originally modified the daemon class to stay around for
testing purposes, it doesn't seem that that is necessary for the
tests to work.

I have thus modified the daemon class to kill its child by default, but
if it turns out that I am wrong and it is necessary to keep the parent
process around when testing, that behavior can be disabled easily.

This helps during actual use, because now a defunct zombie process is
not left lying around when running `jobmon` in daemon mode.