Added an option to leave the parent when daemonizing.
This is important because a few scripts I use run jobmon, and are
intended to continue running after starting the daemon. With the
previous commit, they would die and leave jobmon running in the
background - now, after they were fixed, they stick around.

All this commit does, is add an option which allows scripts calling out
to jobmon (in Python) to either daemonize (and kill themselves) or not
daemonize (and thus keep running).