Add explicit handling for socket timeouts
Since socket timeouts are useful in the command server, but not much elsewhere,
timeouts are now an explicit part of the protocol wrappers so that their timeout
behavior can be configured.

The pipes and sockets used by the event server do not have a timeout, since they
will typically have long wait times between changes. The command server enables
timeouts, so that it can contain misbehaved clients. The status server does not
utilize timeouts for the same reason as the event server.