Attempt to fix issue when starting a service twice
A bug was encountered when doing the following sequence:

    # Process 'foo' crashes
    $ jobmon start foo
    # jobmon tries to restart 'foo' from timer

This occurs because starting the job does not take into account whether
the job will be started again naturally. This fix makes it so that the
user typing `jobmon start foo` will be ignored if there is a pending
restart for foo.